requirements.txt
APScheduler==3.6.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
decorator==5.1.1
idna==3.7
pycparser==2.22
python-dotenv==1.0.1
python-telegram-bot==13.0
pytz==2024.1
requests==2.32.3
setuptools==72.2.0
six==1.16.0
tornado==6.4.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2